Notifications dont render when loaded though they do
